requirements.txt
openai-whisper
keyboard
pyaudio
colorama